split up code into separate files/dirs
